[fix] 미션인증 시 팜클럽 아이디를 받아서 유저팜클럽을 불러오는 형식으로 코드 수정

+++ src/main/java/com/modernfarmer/farmusspring/domain/history/vo/MissionPostHistoryVo.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 
 @Builder(access = AccessLevel.PRIVATE)
 public record MissionPostHistoryVo(
-        Long stepNum,
+        int stepNum,
         String image,
         String content,
         String date
@@ -17,7 +17,7 @@ public record MissionPostHistoryVo(
     public MissionPostHistoryVo {
     }
 
-    public static MissionPostHistoryVo of(Long stepNum, String image, String content, String date) {
+    public static MissionPostHistoryVo of(int stepNum, String image, String content, String date) {
         return MissionPostHistoryVo.builder()
                 .stepNum(stepNum)
                 .image(image)
